Classify benchmarks into micro and macro.
Sandmark now contains lots of tests, many of which are there to stress
test individual language features. Although they might run for a
resonable amount of time and may even allocate lots of memory, they are
not realistic programs.

This commit classifies the benchmarks into macro and micro. The idea
being that macro is realistic and micro isn't. For parallel benchmarks,
macro benchmarks are expected to scale with increasing number of
domains.