Add support for collecting information about GC pausetimes on trunk
Trunk OCaml has support for instrumented runtime, which emits a timing
log. This PR processes that log for each of the benchmark and outputs a
`.bench` file where each line is of the form:

<bench>:{'mean_latency': <mean>,
         'max_latency': <max>,
	 'distr_latency': [<10%>, <20%>, <30%>, <40%>, <50%>, <60%>, <70%>,
	                   <80%>, <90%>, <95%>, <99%>, <99.9%>] }

where the value <99.9%> indicates that 99.9% of the pauses were under
<99.9%> nanoseconds. The unit for the latencies is nanoseconds.

This will only work with instrumented runtime. There is a new compiler
configuration `ocaml-versions/4.11.0+trunk+instrumented.comp` which
enables instrumented runtime by default. This instrumented runtime
should not be used for throughput measurements due to the overheads of
instrumentation.