Fix normalise error when different number of benchmarks in baseline
When displaying plots for the sequential benchmarks, we normalise the metrics
based on the baseline values. The code in `normalise` fails with an error when
the list of benchmarks in the baseline differ from the benchmarks in the
variants that are being compared against the baseline. This commit cleans up
the code to use `pd.DataFrame.pivot` and `pd.DataFrame.melt` to fix this error.