Remove gnuplot dependency for benchmarks
`gnuplot` was added as a dependecy as a part of the PR #343 since a library for
running the Irmin benchmarks depended on it. But, we don't currently run the
`irmin` benchmarks and the `gnuplot` dependency installs a whole bunch of X11
dependencies when setting up the OS for running the benchmarks. This makes the
setup longer than it needs to be, as of now.