Run the run_all_custom.sh script wrapped in flock
The nightly cron jobs on Navajo have started to take almost a day, and
sometimes this could lead to multiple benchmarks getting started at the same
time. To prevent this we use the `flock` utility to detect if the process that
previously owned a lock file is still alive, and not start another process if
that's the case.

The call to `flock` is added to the nightly_* scripts so that it's more visible
to the developers, rather than being hidden away in the crontab file.