Add a ZDD benchmark
This patch adds a benchmark that consists of a simple implementation
of ZDDs (Reduced Ordered Zero-Suppressed Binary Decision Diagrams)
with a small client use that builds a dictionary from a file
containing a word list and searches it using a simple query. This
implementation is likely not a good way to solve the client's search
problem, but it exercises the ZDD implementation.

The implementation uses Weak sets to "hash-cons" the decision diagram
nodes, and uses Ephemerons to memoize the ZDD operations. This might
be interesting as a self-contained benchmark for those features. This
code is not quite a micro-benchmark, but exercises Weak sets and
Ephemerons heavily in a way that may be representative of using them
for hash-consing and memoization.

Signed-off-by: Josh Berdine <josh@berdine.net>