Don't unset/reset OCAML_CONFIG_OPTION and OCAML_RUN_PARAM env vars
The Makefile tries to read OCAML_RUN_PARAM and OCAML_CONFIG_OPTION parameters
from the ocaml-version/*.json files.  This is useful for OCaml compiler
developers when testing their own branches locally, etc, to be able to specify
their own flags for compiling the compiler, etc. But, this unsets the values
that are being set by the `run_all_custom.sh` script using the values specified
in the `sandmark-nightly-config` JSON files.

This commit fixes that to only read these variables from the
ocaml-versions/*.json files if the corresponding environment variables have not
been set, or are empty.